requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
et-xmlfile==2.0.0
h11==0.14.0
idna==3.10
loguru==0.7.3
lxml==5.3.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pysocks==1.7.1
python-dotenv==1.1.0
requests==2.32.3
selenium==4.30.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.29.0
trio-websocket==0.12.2
typing-extensions==4.13.1
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
